+++ src/http/authentication/cert_auth.h
@@ -6,6 +6,7 @@
 #include "crypto/pem.h"
 #include "crypto/verifier.h"
 #include "service/blit.h"
+#include "service/table_names.h"
 #include "service/tables/members.h"
 #include "service/tables/nodes.h"
 #include "service/tables/users.h"
