+++ src/kv/deserialise.h
@@ -6,7 +6,7 @@
 #include "consensus/aft/request.h"
 #include "kv/committable_tx.h"
 #include "kv_types.h"
-#include "service/network_tables.h"
+#include "service/table_names.h"
 
 #include <vector>
