+++ src/kv/store.h
@@ -8,7 +8,6 @@
 #include "kv/committable_tx.h"
 #include "kv_serialiser.h"
 #include "kv_types.h"
-#include "service/network_tables.h"
 #include "snapshot.h"
 
 #define FMT_HEADER_ONLY
